TbfHeaderTlv and expandalbe # of permission bytes

+++ capsules/src/driver.rs
@@ -4,6 +4,11 @@ use enum_primitive::enum_from_primitive;
 enum_from_primitive! {
 #[derive(Debug, PartialEq)]
 // syscall driver numbers
+// when adding a driver, make sure to:
+//   - add the driver to `get_permissions_bit()` below
+//   - add the driver to `self.permission_bits` in tockloader/tockloader/tbfh.py
+//   - increment `NUM_DRIVERS` in tock/kernel/src/tbfheader.rs
+//   - increment `NUM_DRIVERS` in elf2tab/header.rs
 pub enum NUM {
     ADC = 0x00000005,
     ALARM = 0x00000000,
@@ -40,42 +45,40 @@ pub enum NUM {
 
 // sorted in order of ascending driver number
 // this is so it's easier to avoid conflicts when adding a new driver
-pub fn get_permission_bit(driver_num: usize) -> Option<i8> {
+pub fn get_permission_bit(driver_num: usize) -> Option<usize> {
     if let Some(num) = NUM::from_usize(driver_num) {
-        Some(
-            (match num {
-                NUM::ALARM => 0,
-                NUM::CONSOLE => 1,
-                NUM::LED => 2,
-                NUM::BUTTON => 3,
-                NUM::GPIO => 4,
-                NUM::ADC => 5,
-                NUM::DAC => 6,
-                NUM::ANALOG_COMPARATOR => 7,
-                NUM::SPI => 8,
-                NUM::USB_USER => 9,
-                NUM::I2C_MASTER_SLAVE => 10,
-                NUM::BLE_ADVERTISING => 11,
-                NUM::RNG => 12,
-                NUM::CRC => 13,
-                NUM::I2C_MASTER => 14,
-                NUM::APP_FLASH => 15,
-                NUM::NVM_STORAGE => 16,
-                NUM::SD_CARD => 17,
-                NUM::TEMPERATURE => 18,
-                NUM::HUMIDITY => 19,
-                NUM::AMBIENT_LIGHT => 20,
-                NUM::NINEDOF => 21,
-                NUM::TSL2561 => 22,
-                NUM::TMP006 => 23,
-                NUM::LPS25HB => 24,
-                NUM::LTC294X => 25,
-                NUM::MAX17205 => 26,
-                NUM::PCA9544A => 27,
-                NUM::GPIO_ASYNC => 28,
-                NUM::NRF51822_SERIALIZATION => 29,
-            }) as i8,
-        )
+        Some(match num {
+            NUM::ALARM => 0,
+            NUM::CONSOLE => 1,
+            NUM::LED => 2,
+            NUM::BUTTON => 3,
+            NUM::GPIO => 4,
+            NUM::ADC => 5,
+            NUM::DAC => 6,
+            NUM::ANALOG_COMPARATOR => 7,
+            NUM::SPI => 8,
+            NUM::USB_USER => 9,
+            NUM::I2C_MASTER_SLAVE => 10,
+            NUM::BLE_ADVERTISING => 11,
+            NUM::RNG => 12,
+            NUM::CRC => 13,
+            NUM::I2C_MASTER => 14,
+            NUM::APP_FLASH => 15,
+            NUM::NVM_STORAGE => 16,
+            NUM::SD_CARD => 17,
+            NUM::TEMPERATURE => 18,
+            NUM::HUMIDITY => 19,
+            NUM::AMBIENT_LIGHT => 20,
+            NUM::NINEDOF => 21,
+            NUM::TSL2561 => 22,
+            NUM::TMP006 => 23,
+            NUM::LPS25HB => 24,
+            NUM::LTC294X => 25,
+            NUM::MAX17205 => 26,
+            NUM::PCA9544A => 27,
+            NUM::GPIO_ASYNC => 28,
+            NUM::NRF51822_SERIALIZATION => 29,
+        })
     } else {
         None
     }
